DAL全部完成替换
TODO:
1. RepairDetailPage 添加上传照片

+++ DataAccess/Entity/projectDetailTable.cs
@@ -36,6 +36,6 @@ namespace DataAccess.Entity
 
         public DateTime? updateDate { get; set; }
 
-        public byte deleteStatus { get; set; }
+        public int deleteStatus { get; set; }
     }
 }
